cmd: Add parse options and teach "--line-numbers"
Adding the ability to parse command line options using the libc getopt.h
library in order to make it easier to parse more command options. Both
shorter and longer command are supported.

To simplify the development a parse_opts() are implemented that will
reteurn an allocated struct with the available options.

Teach findcode to take a `n` or `--line-numbers` as argument to tune
whether to show the line numbers where the block started and ended.