Introduce mmap() to map file into memory
The findcode process_file() uses open a file stream and utilize
the `fread()` in order to read the bytes from the disk. This also
required the appropriate memory are allocated in order to store
the data returned by the function.

According to research, the `mmap()` provides a better interface
as the hard work are done by the kernel OS and the kernel does
not need to copy data within the user process, kernel process
and DMA.

Replace `mmap()` to map the file into memory instead of reading
the file from a disk stream in order to improve the performance
and robutness (aka: let the work be done by the kernel)