close fd after mmap in process_file()
After the file is mmap'ed into memory, there is not need to
keep the file descriptor (fd) open. Close the file right
after the mmap() function is called so we close the file
as soon as we are done with it.

Additionally, two close() function are removed as it become
the normal code path to close it.